Changed js and scss folder structure
