+++ etsin_finder/frontend/.eslintrc
@@ -4,10 +4,11 @@
     "react"
   ],
   "env": {
-    "browser": true
+    "browser": true,
+    "jest": true,
   },
   "rules": {
-//    "max-len": ["error", { "code": 130 }],
+    "max-len": ["error", { "code": 150 }],
     "no-extra-semi": 1,
     "semi": 0,
     "react/prop-types": 0,
