+++ etsin_finder/frontend/js/app.jsx
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { Provider } from 'mobx-react';
 
-import ErrorBoundary from './components/errorBoundary'
+import ErrorBoundary from './components/general/errorBoundary'
 import Footer from './layout/footer';
 import Header from './layout/header';
 import Content from './layout/content';
