+++ etsin_finder/frontend/js/components/dataset/Sidebar.jsx
@@ -1,15 +1,15 @@
 import React, { Component } from 'react'
 import Translate from 'react-translate-component'
-import DsSidebarItem from './dsSidebarItem'
-import Locale from '../stores/view/language'
-import checkNested from '../utils/checkNested'
-import DateFormat from './dateFormat'
-import Identifier from './identifier'
-import ErrorBoundary from './errorBoundary'
+import SidebarItem from './SidebarItem'
+import Locale from '../../stores/view/language'
+import checkNested from '../../utils/checkNested'
+import DateFormat from './data/dateFormat'
+import Identifier from './data/identifier'
+import ErrorBoundary from '../general/errorBoundary'
 
-import checkDataLang from '../utils/checkDataLang'
+import checkDataLang from '../../utils/checkDataLang'
 
-export default class DsSidebar extends Component {
+export default class Sidebar extends Component {
   dateSeparator(start, end) {
     return start && end
       ? <p key={start}><DateFormat date={start} /> - <DateFormat date={end} /></p>
@@ -26,63 +26,63 @@ export default class DsSidebar extends Component {
       <div className="sidebar content-box">
         <div className="separator">
           <ErrorBoundary>
-            <DsSidebarItem component="p" trans="dataset.publisher" fallback="Publisher">
+            <SidebarItem component="p" trans="dataset.publisher" fallback="Publisher">
               {
                 checkNested(researchDataset, 'publisher', 'name')
                 ? researchDataset.publisher.name[currentLang]
                 : null
               }
-            </DsSidebarItem>
+            </SidebarItem>
           </ErrorBoundary>
         </div>
         <div className="separator">
           <ErrorBoundary>
-            <DsSidebarItem component="p" trans="dataset.identifier" fallback="Identifier">
+            <SidebarItem component="p" trans="dataset.identifier" fallback="Identifier">
               <Identifier idn={researchDataset.preferred_identifier}>
                 {researchDataset.preferred_identifier}
               </Identifier>
-            </DsSidebarItem>
+            </SidebarItem>
           </ErrorBoundary>
         </div>
         <div>
           <ErrorBoundary>
-            <DsSidebarItem component="p" trans="dataset.project" fallback="Project" hideEmpty="true">
+            <SidebarItem component="p" trans="dataset.project" fallback="Project" hideEmpty="true">
               {
                 isOutputOf
                   ? researchDataset.is_output_of.map(item => item.name[currentLang])
                   : null
               }
-            </DsSidebarItem>
+            </SidebarItem>
           </ErrorBoundary>
           <ErrorBoundary>
-            <DsSidebarItem component="div" trans="dataset.field_of_science" fallback="Field of Science" hideEmpty="true">
+            <SidebarItem component="div" trans="dataset.field_of_science" fallback="Field of Science" hideEmpty="true">
               {dataCatalog.catalog_json.field_of_science.map(field => (
                 <p key={field.identifier}>
                   {field.pref_label[this.props.lang]}
                 </p>
               ))}
-            </DsSidebarItem>
+            </SidebarItem>
           </ErrorBoundary>
           <ErrorBoundary>
-            <DsSidebarItem component="p" trans="dataset.keywords" fallback="Keywords" hideEmpty="true">
+            <SidebarItem component="p" trans="dataset.keywords" fallback="Keywords" hideEmpty="true">
               {
                 researchDataset.keyword
                   ? researchDataset.keyword.map(keyword => <span className="keyword" key={keyword}>{keyword} </span>)
                   : null
               }
-            </DsSidebarItem>
+            </SidebarItem>
           </ErrorBoundary>
           <ErrorBoundary>
-            <DsSidebarItem component="p" trans="dataset.spatial_coverage" fallback="Spatial Coverage" hideEmpty="true">
+            <SidebarItem component="p" trans="dataset.spatial_coverage" fallback="Spatial Coverage" hideEmpty="true">
               {
                 checkNested(researchDataset, 'spatial', 'geographic_name')
                   ? researchDataset.spatial.geographic_name
                   : null
               }
-            </DsSidebarItem>
+            </SidebarItem>
           </ErrorBoundary>
           <ErrorBoundary>
-            <DsSidebarItem component="div" trans="dataset.temporal_coverage" fallback="Temporal Coverage" hideEmpty="true">
+            <SidebarItem component="div" trans="dataset.temporal_coverage" fallback="Temporal Coverage" hideEmpty="true">
               {
                 checkNested(researchDataset, 'temporal')
                   ? researchDataset.temporal.map(dates => this.dateSeparator(
@@ -91,10 +91,10 @@ export default class DsSidebar extends Component {
                     ))
                   : null
               }
-            </DsSidebarItem>
+            </SidebarItem>
           </ErrorBoundary>
           <ErrorBoundary>
-            <DsSidebarItem component="div" trans="dataset.license" fallback="License" hideEmpty="true">
+            <SidebarItem component="div" trans="dataset.license" fallback="License" hideEmpty="true">
               {
                 checkNested(researchDataset, 'access_rights', 'license')
                   ? researchDataset.access_rights.license.map(rights => (
@@ -104,10 +104,10 @@ export default class DsSidebar extends Component {
                   ))
                   : null
               }
-            </DsSidebarItem>
+            </SidebarItem>
           </ErrorBoundary>
           <ErrorBoundary>
-            <DsSidebarItem component="p" trans="dataset.access_rights" fallback="Access rights statement" hideEmpty="true">
+            <SidebarItem component="p" trans="dataset.access_rights" fallback="Access rights statement" hideEmpty="true">
               {
                 checkNested(dataCatalog, 'catalog_json', 'access_rights', 'description')
                   ? checkDataLang(
@@ -116,10 +116,10 @@ export default class DsSidebar extends Component {
                   )
                   : null
               }
-            </DsSidebarItem>
+            </SidebarItem>
           </ErrorBoundary>
           <ErrorBoundary>
-            <DsSidebarItem component="p" trans="dataset.funder" fallback="Funder" hideEmpty="true">
+            <SidebarItem component="p" trans="dataset.funder" fallback="Funder" hideEmpty="true">
               {
                 checkNested(researchDataset, 'is_output_of')
                   ? researchDataset.is_output_of.map(output => (
@@ -131,31 +131,31 @@ export default class DsSidebar extends Component {
                   ))
                   : null
               }
-            </DsSidebarItem>
+            </SidebarItem>
           </ErrorBoundary>
           <ErrorBoundary>
-            <DsSidebarItem component="p" trans="dataset.curator" fallback="Curator" hideEmpty="true">
+            <SidebarItem component="p" trans="dataset.curator" fallback="Curator" hideEmpty="true">
               {researchDataset.curator ? researchDataset.curator.map(curators => (
                 checkDataLang(curators.name, currentLang)
               )) : null}
-            </DsSidebarItem>
+            </SidebarItem>
           </ErrorBoundary>
           <ErrorBoundary>
-            <DsSidebarItem component="p" trans="dataset.infrastructure" fallback="Infrastructure" hideEmpty="true">
+            <SidebarItem component="p" trans="dataset.infrastructure" fallback="Infrastructure" hideEmpty="true">
               {
                 checkNested(researchDataset, 'related_entity')
                   ? researchDataset.related_entity.map(entity => (
                     checkDataLang(entity.title, currentLang)))
                   : null
               }
-            </DsSidebarItem>
+            </SidebarItem>
           </ErrorBoundary>
           <ErrorBoundary>
-            <DsSidebarItem component="p" trans="dataset.permanent_link" fallback="Permanent link to this page" hideEmpty="true">
+            <SidebarItem component="p" trans="dataset.permanent_link" fallback="Permanent link to this page" hideEmpty="true">
               {
                 `something/something/${researchDataset.urn_identifier}`
               }
-            </DsSidebarItem>
+            </SidebarItem>
           </ErrorBoundary>
         </div>
         <div className="row no-gutters justify-content-md-center">
