+++ etsin_finder/frontend/js/components/dataset/Tabs.jsx
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import Translate from 'react-translate-component'
 import { NavLink } from 'react-router-dom'
 
-export default class DsTabs extends Component {
+export default class Tabs extends Component {
   constructor(props) {
     super(props)
     console.log(this.props.identifier)
