+++ etsin_finder/frontend/js/components/general/hero.jsx
