+++ etsin_finder/frontend/js/components/search/filterResults/index.jsx
