+++ etsin_finder/frontend/scss/components/dataset/_dsContent.scss
