+++ etsin_finder/frontend/scss/components/dataset/_dsSidebar.scss
