+++ etsin_finder/frontend/scss/components/search/_filter.scss
