+++ etsin_finder/frontend/scss/components/search/_listItem.scss
