+++ etsin_finder/frontend/scss/components/search/_searchBar.scss
